+++ app/routes/submit.js
@@ -25,9 +25,10 @@ export default Route.extend(ConfirmationMixin, ResetScrollMixin, CasAuthenticate
     }),
     model() {
         // Store the empty preprint to be created on the model hook for page. Node will be fetched
-        //  internally during submission process.
+        // internally during submission process.
         return this.get('store').createRecord('preprint', {
             subjects: [],
+            title: '',
         });
     },
     afterModel() {
@@ -35,6 +36,7 @@ export default Route.extend(ConfirmationMixin, ResetScrollMixin, CasAuthenticate
     },
     setupController(controller, model) {
         this.setupSubmitController(controller, model);
+        controller._setCurrentProvider();
         return this._super(...arguments);
     },
     isPageDirty() {